+++ deploy/build.sh
@@ -1,7 +1,9 @@
 #!/usr/bin/env bash
 
 # Usage: build.sh [<work-dir>]
-cd ${1:-"."}
+cd ${1:-"."} || exit 1
+
+mkdir -p var/{panel-assets,panel-assets-history}
 
 composer install --no-dev --no-interaction
 if [ $? -ne 0 ]; then exit 1; fi