+++ src/Entity/PanelAsset.php
@@ -2,32 +2,34 @@
 
 namespace App\Entity;
 
-use App\Repository\AssetRepository;
+use App\Repository\PanelAssetRepository;
 use DateTimeImmutable;
-use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
 use Gedmo\Mapping\Annotation as Gedmo;
 
-#[ORM\Entity(repositoryClass: AssetRepository::class)]
-#[ORM\UniqueConstraint(name: 'UNIQUE_NAME', columns: ['module', 'name'])]
-class Asset
+#[ORM\Entity(repositoryClass: PanelAssetRepository::class)]
+#[ORM\UniqueConstraint(name: 'UNIQUE_FILE_ID', columns: ['file_id'])]
+#[Gedmo\Loggable]
+class PanelAsset
 {
     #[ORM\Id]
     #[ORM\GeneratedValue]
     #[ORM\Column]
     private ?int $id = null;
 
-    #[ORM\Column(length: 31)]
-    private ?string $module = null;
+    #[ORM\Column(name: 'file_id', length: 255)]
+    #[Gedmo\Versioned]
+    private ?string $fileId = null;
 
-    #[ORM\Column(length: 255)]
-    private ?string $name = null;
+    #[ORM\Column(name: 'file_path', length: 255)]
+    private ?string $filePath = null;
 
     #[ORM\Column(name: 'mime_type', length: 127)]
     private ?string $mimeType = null;
 
-    #[ORM\Column(type: Types::BLOB)]
-    private $content = null;
+    #[ORM\Column(length: 16)]
+    #[Gedmo\Versioned]
+    private ?string $hash = null;
 
     #[ORM\Column(name: 'created_at')]
     #[Gedmo\Timestampable(on: 'create')]
@@ -42,25 +44,36 @@ class Asset
         return $this->id;
     }
 
-    public function getModule(): ?string
+    public function getFileId(): ?string
     {
-        return $this->module;
+        return $this->fileId;
     }
 
-    public function setModule(string $module): self
+    public function setFileId(string $fileId): self
     {
-        $this->module = $module;
+        $this->fileId = $fileId;
         return $this;
     }
 
-    public function getName(): ?string
+    public function getFileName(): ?string
     {
-        return $this->name;
+        if (null === $this->filePath) {
+            return null;
+        }
+        if (false !== ($index = strrpos($this->filePath, '/'))) {
+            return substr($this->filePath, $index + 1);
+        }
+        return $this->filePath;
     }
 
-    public function setName(string $name): self
+    public function getFilePath(): ?string
     {
-        $this->name = $name;
+        return $this->filePath;
+    }
+
+    public function setFilePath(string $filePath): self
+    {
+        $this->filePath = $filePath;
         return $this;
     }
 
@@ -75,14 +88,15 @@ class Asset
         return $this;
     }
 
-    public function getContent()
+    public function getHash(): ?string
     {
-        return $this->content;
+        return $this->hash;
     }
 
-    public function setContent($content): self
+    public function setHash(string $hash): self
     {
-        $this->content = $content;
+        $this->hash = $hash;
+
         return $this;
     }