+++ src/Messenger/Message/AdPanelReload.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Messenger\Message;
+
+final class AdPanelReload
+{
+}