+++ src/Messenger/Message/AdServerCommand.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Messenger\Message;
+
+interface AdServerCommand
+{
+    public function getSignature(): string;
+
+    public function getArguments(): array;
+}