+++ src/Messenger/Message/AdServerFetchExchangeRate.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Messenger\Message;
+
+final class AdServerFetchExchangeRate implements AdServerCommand
+{
+    public function getSignature(): string
+    {
+        return 'ops:exchange-rate:fetch';
+    }
+
+    public function getArguments(): array
+    {
+        return [];
+    }
+}