+++ src/Messenger/Message/AdServerUpdateSiteRank.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Messenger\Message;
+
+final class AdServerUpdateSiteRank implements AdServerCommand
+{
+    public function getSignature(): string
+    {
+        return 'ops:supply:site-rank:update';
+    }
+
+    public function getArguments(): array
+    {
+        return ['--all'];
+    }
+}