+++ src/Messenger/Message/AdServerUpdateTargeting.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Messenger\Message;
+
+final class AdServerUpdateTargeting implements AdServerCommand
+{
+    public function getSignature(): string
+    {
+        return 'ops:targeting-options:update';
+    }
+
+    public function getArguments(): array
+    {
+        return [];
+    }
+}