+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2022-10-26
+### Added
+- Custom assets
+- Custom CSS
+- Reload AdPanel asynchronously
+- Check upload file limit
+
 ## [1.0.2] - 2022-10-04
 ### Added
 - SMTP configuration
@@ -18,18 +25,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2022-09-28
 ### Added
 - Controller
+### Changed
+- ADS account checker verifies balance
 ### Fixed
 - Send partial data in installer's steps
 - Send installer's step data if required
-### Changed
-- ADS account checker verifies balance
 
 ## [0.1.0] - 2022-08-12
 ### Added
 - Installer
 - Project setup
 
-[Unreleased]: https://github.com/adshares/adcontroller/compare/v1.0.2...develop
+[Unreleased]: https://github.com/adshares/adcontroller/compare/v1.1.0...develop
+[1.1.0]: https://github.com/adshares/adcontroller/compare/v1.0.2...v1.1.0
 [1.0.2]: https://github.com/adshares/adcontroller/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/adshares/adcontroller/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/adshares/adcontroller/compare/v0.1.0...v1.0.0